+++ R/utils-xml.R
@@ -1,6 +1,8 @@
 fix_nodes <- function(nodes = NULL) {
   if (length(nodes) == 0) return(nodes)
+  translate_overview(nodes)
   fix_headings(nodes)
+  fix_accordions(nodes)
   fix_callouts(nodes)
   fix_codeblocks(nodes)
   fix_figures(nodes)
@@ -21,18 +23,64 @@ fix_headings <- function(nodes = NULL) {
   invisible(nodes)
 }
 
+#' add codewrapper class and apply code heading to all code blocks
+#'
+#' The syntax highlighte4d code blocks that come out of pandoc have this
+#' structure (where `lang` is the language of the code block):
+#'
+#' ```html
+#' <div class="sourceCode" id="cb1">
+#'   <pre class="sourceCode lang">
+#'     <code class="sourceCode lang">
+#'     ...
+#'     </code>
+#'   </pre>
+#' </div>
+#' ```
+#'
+#' In The Workbench, we want to have this structure:
+#'
+#' ```html
+#' <div class="codewrapper sourceCode" id="cb1">
+#'   <h3 class="code-label">
+#'    LANG
+#'    <i aria-hidden=true data-feather="chevron-left"></i>
+#'    <i aria-hidden=true data-feather="chevron-right"></i>
+#'   </h3>
+#'   <pre class="sourceCode lang" tabindex="0">
+#'     <code class="sourceCode lang">
+#'     ...
+#'     </code>
+#'   </pre>
+#' </div>
+#' ```
+#'
+#' This allows us to display the language of the code block in the lesson,
+#' which can be helpful when the lesson switches between BASH and another
+#' language.
+#'
+#' @param nodes HTML nodes
+#' @return the modified nodes
+#'
+#' @noRd
 fix_codeblocks <- function(nodes = NULL) {
   if (length(nodes) == 0) return(nodes)
   code <- xml2::xml_find_all(nodes, ".//div[starts-with(@class, 'sourceCode')]")
   xml2::xml_set_attr(code, "class", "codewrapper sourceCode")
   pre <- xml2::xml_children(code)
-  type <- rev(trimws(sub("sourceCode", "", xml2::xml_attr(pre, "class"))))
-  add_code_heading(pre, toupper(type))
-  outputs <- xml2::xml_find_all(nodes, ".//pre[@class='output' or @class='warning' or @class='error']")
+  # pre-compile these during the transformation so we only have to do it
+  # once per document
+  translations <- get_codeblock_translations()
+  # Extract the language, transform to all caps, and reverse the order.
+  # We need to reverse the order so that we can add
+  type <- toupper(trimws(sub("sourceCode", "", xml2::xml_attr(pre, "class"))))
+  add_code_heading(pre, apply_translations(type, translations))
+  outputs <- xml2::xml_find_all(nodes,
+    ".//pre[@class='output' or @class='warning' or @class='error']")
   if (length(outputs)) {
     xml2::xml_add_parent(outputs, "div", class = "codewrapper")
-    class_headings <- rev(toupper(xml2::xml_attr(outputs, "class")))
-    add_code_heading(outputs, class_headings)
+    class_headings <- toupper(trimws(xml2::xml_attr(outputs, "class")))
+    add_code_heading(outputs, apply_translations(class_headings, translations))
   }
   invisible(nodes)
 }
@@ -40,8 +88,11 @@ fix_codeblocks <- function(nodes = NULL) {
 add_code_heading <- function(codes = NULL, labels = "OUTPUT") {
   if (length(codes) == 0) return(codes)
   xml2::xml_set_attr(codes, "tabindex", "0")
-  heads <- xml2::xml_add_sibling(codes, "h3", labels, class = "code-label",
-    .where = "before")
+  # NOTE: xml_add_sibling adds the siblings from bottom to top, so these labels
+  # need to be in reverse. It's weird.
+  heads <- xml2::xml_add_sibling(codes, "h3", rev(labels),
+    class = "code-label", .where = "before"
+  )
   for (head in heads) {
     xml2::xml_add_child(head, "i",
       "aria-hidden" = "true", "data-feather" = "chevron-left")
@@ -80,7 +131,7 @@ add_class <- function(nodes, new) {
 
 add_anchors <- function(nodes, ids) {
   anchor <- paste0(
-    "<a class='anchor' aria-label='anchor' href='#", ids, "'></a>"
+    "<a class='anchor' aria-label='", tr_("anchor"), "' href='#", ids, "'></a>"
   )
   for (i in seq_along(nodes)) {
     heading <- nodes[[i]]
@@ -93,10 +144,61 @@ add_anchors <- function(nodes, ids) {
   }
 }
 
+# translate the overview cards, which are defined in
+# inst/rmarkdown/lua/lesson.lua
+translate_overview <- function(nodes = NULL) {
+  if (length(nodes) == 0) return(nodes)
+  card <- xml2::xml_find_first(nodes, ".//div[@class='overview card']")
+  if (length(card) == 0) {
+    return(nodes)
+  }
+  overview <- xml2::xml_find_first(card, "./h2[@class='card-header']")
+  qpath <- ".//div[starts-with(@class, 'inner')]/h3[@class='card-title'][text()='Questions']"
+  opath <- ".//div[starts-with(@class, 'inner')]/h3[@class='card-title'][text()='Objectives']"
+  questions <- xml2::xml_find_first(card, qpath)
+  objectives <- xml2::xml_find_first(card, opath)
+
+  xml2::xml_set_text(questions, tr_("Questions"))
+  xml2::xml_set_text(objectives, tr_("Objectives"))
+  xml2::xml_set_text(overview, tr_("Overview"))
+  invisible(nodes)
+}
+
+# translate contents of an XML node list
+# @param nodes an xml node or xml nodelist
+# @param translations a named vector of translated strings whose names are the
+#   strings in English
+xml_text_translate <- function(nodes, translations) {
+  txt <- xml2::xml_text(nodes, trim = TRUE)
+  xml2::xml_set_text(nodes, apply_translations(txt, translations))
+  return(invisible(nodes))
+}
+
+fix_accordions <- function(nodes = NULL) {
+  if (length(nodes) == 0) return(nodes)
+  accordions <- xml2::xml_find_all(nodes,
+    ".//div[starts-with(@class, 'accordion ')]"
+  )
+  # NOTE: we need to include `text()` in the call here because of the presence
+  # of the decorative blocks inside the accordion headings.
+  # solution and hint are h4
+  # instructor and spoiler are h3
+  headings <- xml2::xml_find_all(accordions,
+    "./div/button/h3/text() | ./div/button/h4/text()"
+  )
+  translations <- get_accordion_translations()
+  xml_text_translate(headings, translations)
+  # at this point, we would fix headings, but we do not actually have a way to
+  # consistently do this, so it remains as an exercise for the future.
+  return(invisible(nodes))
+}
+
 fix_callouts <- function(nodes = NULL) {
   if (length(nodes) == 0) return(nodes)
   callouts <- xml2::xml_find_all(nodes, ".//div[starts-with(@class, 'callout ')]")
   h3 <- xml2::xml_find_all(callouts, "./div/h3")
+  translations <- get_callout_translations()
+  xml_text_translate(h3, translations)
   xml2::xml_set_attr(h3, "class", "callout-title")
   inner_div <- xml2::xml_parent(h3)
   # remove the "section level3 callout-title" attrs